--- pkg/root/root.go
@@ -1,11 +1,0 @@
-package root
-
-type BaseRequest struct {
-	Url     string
-	Headers []Header
-}
-
-type Header struct {
-	Name  string
-	Value string
-}
